+++ common/controllers/modulelifecycle/manager.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2023, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 
-package lifecycle
+package modulelifecycle
 
 import (
 	"github.com/verrazzano/verrazzano-modules/common/controllers/base/basecontroller"
