+++ common/lifecycle-actions/action_spi/action_spi.go
